app/requirements.txt
fastapi
pydantic
requests
python-dotenv
uvicorn
google.cloud
jupyter
nbformat
nbconvert